requirements.txt
# Core LangChain dependencies
langchain==0.3.26
langchain-core==0.3.68
langchain-community==0.3.27
langchain-anthropic==0.3.2

# Claude API client
anthropic>=0.8.0

# Environment variables
python-dotenv>=0.19.0

# RAG evaluation framework
ragas==0.2.15

# ML and data processing
datasets==4.0.0
numpy==2.3.1
pandas==2.3.1

# Vector search
faiss-cpu==1.11.0

# Embeddings (HuggingFace)
sentence-transformers==3.3.1
transformers==4.46.3
torch==2.5.1

# Evaluation metrics dependencies
rouge_score==0.1.2
sacrebleu==2.5.1

# Additional dependencies for text processing
nltk==3.9.1
tiktoken==0.9.0

# HTTP and API clients
requests==2.32.4
httpx==0.28.1

# Data serialization
PyYAML==6.0.2